requirements.txt
requests
openai
pydantic_core
fastapi==0.99.1
PyGithub
